+++ src/fintrist_ds/settings.py
@@ -10,7 +10,8 @@ import fintrist
 load_dotenv()
 
 class ConfigObj():
-    APIKEY = os.getenv('APIKEY')
+    APIKEY_AV = os.getenv('APIKEY_AV')
+    APIKEY_TIINGO = os.getenv('APIKEY_TIINGO')
     DASK_PORT = int(os.getenv('DASK_PORT') or 8786)
     NUM_WORKERS = int(os.getenv('NUM_WORKERS') or 4)
     DASK_HOST = f"{fintrist.Config.DB_HOST}:{DASK_PORT}"